requirements.txt
PyQt5>=5.15,<6
Pillow>=10,<12
torch>=2.3,<3
torchvision>=0.18,<1
onnx>=1.16,<2
